+++ src/__mocks__/vscode.ts
@@ -37,6 +37,20 @@ export class CodeActionKind {
   constructor(readonly value: string) {}
 }
 
+export enum CodeActionTriggerKind {
+  Invoke = 1,
+  Automatic = 2,
+}
+
+export class CompletionList {
+  constructor(readonly items: any[]) {}
+}
+
+export enum CompletionItemKind {
+  EnumMember,
+  Property,
+}
+
 export class Diagnostic {
   code?: string
   source?: string
@@ -69,9 +83,26 @@ export const ProgressLocation = {
 
 export class Range {
   constructor(readonly start: Position, readonly end: Position) {}
+
+  contains(position: Position) {
+    return !(
+      position.line < this.start.line ||
+      position.line > this.end.line ||
+      (position.line === this.start.line &&
+        position.character < this.start.character) ||
+      (position.line === this.end.line &&
+        position.character > this.end.character)
+    )
+  }
+
+  with(start?: Position, end?: Position) {
+    return new Range(start ?? this.start, end ?? this.end)
+  }
 }
 
 export const RelativePattern = jest.fn()
+
+export const ThemeColor = jest.fn(() => '#000000ff')
 
 export const Uri = {
   file: uriInstance,
@@ -104,6 +135,10 @@ export enum FileType {
   SymbolicLink = 64,
 }
 
+export class Hover {
+  constructor(public contents: string, public range?: Range) {}
+}
+
 export const languages = {
   createDiagnosticCollection: jest.fn((name) => ({
     name,
@@ -111,7 +146,17 @@ export const languages = {
     set: jest.fn(),
   })),
   registerCodeActionsProvider: jest.fn(),
+  registerCompletionItemProvider: jest.fn(),
+  registerHoverProvider: jest.fn(),
   setTextDocumentLanguage: jest.fn(),
+}
+
+export class MarkdownString {
+  constructor(public value: string) {}
+
+  toString() {
+    return this.value
+  }
 }
 
 export let version: string = defaultVSCodeVersion
@@ -121,6 +166,7 @@ export const _setVSCodeVersion = (value: string) => {
 
 export const window = {
   activeTextEditor: undefined,
+  createTextEditorDecorationType: jest.fn((t) => t),
   onDidChangeActiveTextEditor: jest.fn(),
   showErrorMessage: jest.fn(),
   showQuickPick: jest.fn(),
@@ -142,9 +188,13 @@ export const workspace = {
     ),
   })),
   getWorkspaceFolder: jest.fn(),
+  get isTrusted() {
+    return true
+  },
   onDidChangeConfiguration: jest.fn(),
   onDidChangeTextDocument: jest.fn(),
   onDidCloseTextDocument: jest.fn(),
+  onDidGrantWorkspaceTrust: jest.fn(),
   textDocuments: [] as any,
 
   _setConfiguration: (conf: MockedConf = {}) => {