+++ src/commands/show-quick-pick.ts
@@ -1,12 +1,12 @@
-import { commands, QuickPickItem, window } from 'vscode'
+import { commands, QuickPickItem, window, workspace } from 'vscode'
 import { contributes } from '../../package.json'
 import { command as exportCommand } from './export'
 import { command as openExtensionSettingsCommand } from './open-extension-settings'
-import { command as toggleMarpPreviewCommand } from './toggle-marp-preview'
+import { command as toggleMarpFeatureCommand } from './toggle-marp-feature'
 
 export const cmdSymbol = Symbol()
 
-const contributedCommands = [exportCommand, toggleMarpPreviewCommand]
+const contributedCommands = [exportCommand, toggleMarpFeatureCommand]
 const availableCommands: (QuickPickItem & { [cmdSymbol]: string })[] = []
 
 for (const cmdPath of contributedCommands) {
@@ -28,11 +28,23 @@ availableCommands.push({
 
 export const command = 'markdown.marp.showQuickPick'
 
+const isTrustedCommand = (cmd: string) => {
+  if (!workspace.isTrusted && cmd === exportCommand) return false
+  return true
+}
+
 export default async function showQuickPick() {
-  const command = await window.showQuickPick(availableCommands, {
-    matchOnDescription: true,
-    placeHolder: 'Select available command in Marp for VS Code...',
-  })
+  const command = await window.showQuickPick(
+    availableCommands.map((cmd) =>
+      isTrustedCommand(cmd[cmdSymbol])
+        ? cmd
+        : { ...cmd, description: `${cmd.description} $(shield)` }
+    ),
+    {
+      matchOnDescription: true,
+      placeHolder: 'Select available command in Marp for VS Code...',
+    }
+  )
 
   if (command?.[cmdSymbol]) {
     await commands.executeCommand(command[cmdSymbol])