+++ src/diagnostics/deprecated-dollar-prefix.test.ts
@@ -2,6 +2,7 @@ import dedent from 'dedent'
 import {
   CancellationToken,
   CodeAction,
+  CodeActionTriggerKind,
   CodeActionKind,
   Diagnostic,
   DiagnosticSeverity,
@@ -11,6 +12,7 @@ import {
   TextDocument,
   WorkspaceEdit,
 } from 'vscode'
+import { DirectiveParser } from '../directives/parser'
 import * as rule from './deprecated-dollar-prefix'
 
 jest.mock('vscode')
@@ -29,9 +31,12 @@ const doc = (text: string): TextDocument =>
 
 describe('[Diagnostics rule] Deprecated dollar prefix', () => {
   const register = (doc: TextDocument): Diagnostic[] => {
+    const parser = new DirectiveParser()
     const diagnostics: Diagnostic[] = []
-    rule.register(doc, diagnostics)
 
+    rule.register(doc, parser, diagnostics)
+
+    parser.parse(doc)
     return diagnostics
   }
 
@@ -164,7 +169,7 @@ describe('[Diagnostics rule] Deprecated dollar prefix', () => {
         const codeActions = new rule.RemoveDollarPrefix().provideCodeActions(
           document,
           dummyRange,
-          { diagnostics },
+          { diagnostics, triggerKind: CodeActionTriggerKind.Invoke },
           dummyToken
         )
 
@@ -192,7 +197,7 @@ describe('[Diagnostics rule] Deprecated dollar prefix', () => {
         const codeActions = new rule.RemoveDollarPrefix().provideCodeActions(
           document,
           dummyRange,
-          { diagnostics },
+          { diagnostics, triggerKind: CodeActionTriggerKind.Invoke },
           dummyToken
         )