+++ src/diagnostics/index.ts
@@ -7,15 +7,23 @@ import {
   window,
   workspace,
 } from 'vscode'
+import { DirectiveParser } from '../directives/parser'
 import { detectMarpDocument } from '../utils'
 import * as deprecatedDollarPrefix from './deprecated-dollar-prefix'
+import * as overloadingGlobalDirective from './overloading-global-directive'
+import * as unknownTheme from './unknown-theme'
 
 export const collection = languages.createDiagnosticCollection('marp-vscode')
 
 const setDiagnostics = lodashDebounce((doc: TextDocument) => {
+  const directiveParser = new DirectiveParser()
   const diagnostics: Diagnostic[] = []
 
-  deprecatedDollarPrefix.register(doc, diagnostics)
+  deprecatedDollarPrefix.register(doc, directiveParser, diagnostics)
+  overloadingGlobalDirective.register(doc, directiveParser, diagnostics)
+  unknownTheme.register(doc, directiveParser, diagnostics)
+
+  directiveParser.parse(doc)
 
   collection.set(doc.uri, diagnostics)
 }, 500)