+++ src/workspace-proxy-server.ts
@@ -1,9 +1,9 @@
-import type { Server } from 'http'
 import path from 'path'
 import { URL } from 'url'
 import express from 'express'
 import { getPortPromise } from 'portfinder'
 import { FileType, workspace, WorkspaceFolder } from 'vscode'
+import type { Server } from 'http'
 
 export interface WorkspaceProxyServer {
   dispose: () => void