+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 ## [Unreleased]
 
+### Breaking
+
+- VS Code >= 1.56 is now required ([#231](https://github.com/marp-team/marp-vscode/pull/231))
+- Renamed command `markdown.marp.toggleMarpPreview` to `markdown.marp.toggleMarpFeature` ([#235](https://github.com/marp-team/marp-vscode/pull/235))
+
+### Added
+
+- IntelliSense for Marp directives ([#158](https://github.com/marp-team/marp-vscode/issues/158), [#235](https://github.com/marp-team/marp-vscode/pull/235))
+  - Auto completion for directive keys and some values
+  - Syntax highlight for recognized directive keys
+  - Hover help for recognized directives
+
+* `overloading-global-directive` diagnostic: Mark overloaded global directive in the subsequent definition ([#232](https://github.com/marp-team/marp-vscode/pull/232))
+* `unknown-theme` diagnostic: Mark if the specified theme name is not recognized by the extension ([#236](https://github.com/marp-team/marp-vscode/pull/236))
+
+### Changed
+
+- Support [Workspace Trust](https://code.visualstudio.com/updates/v1_56#_workspace-trust): Restrict some features in the untrusted workspace ([#231](https://github.com/marp-team/marp-vscode/pull/231))
+- Upgrade Marp Core to [v2.0.3](https://github.com/marp-team/marp-core/releases/v2.0.3) ([#237](https://github.com/marp-team/marp-vscode/pull/237))
+- Upgrade Marp CLI to [v1.1.1](https://github.com/marp-team/marp-cli/releases/v1.1.1) ([#237](https://github.com/marp-team/marp-vscode/pull/237))
+  - PPTX export will become to pre-render images with high resolution
+- Upgrade dependent packages to the latest version ([#237](https://github.com/marp-team/marp-vscode/pull/237))
+
 ## v0.19.1 - 2021-05-11
 
 ### Fixed